Refactor proof handlers
